requirements.txt
crewai
streamlit
openai
unstructured
pyowm
langchain
langchain-community
langchain-core
langchain-groq
tools
requests
fastapi
uvicorn
pydantic
python-dotenv
langchain-openai
langchain-google-genai
genai